pyproject.toml
[project]
name = "open-extract"
version = "0.1.0"
description = "Extract structural data from PDFs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.6",
    "ipykernel>=6.29.5",
    "langflow>=1.1.4.post1",
    "llama-parse>=0.5.20",
    "markitdown>=0.0.1a3",
    "matplotlib>=3.10.0",
    "ollama>=0.4.7",
    "openai>=1.61.1",
    "openparse>=0.7.0",
    "pymupdf>=1.25.2",
    "pymupdf4llm>=0.0.17",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
]

[tool.setuptools]
packages = ["open_extract"]
